requirements.txt
flask
flask-bcrypt
flask-session
mysql-connector-python
python-dotenv
